refactored some code

+++ min-food-waste-backend/src/main/java/net/barnowlstudio/food_waste_backend/business/ISpoonacularService.java
@@ -6,5 +6,5 @@ import net.barnowlstudio.food_waste_backend.dto.response.RecipeResponse;
 public interface ISpoonacularService {
     RandomRecipesResponse getRandomRecipes(int numberOfRecipes);
 
-    RecipeResponse getRecipeById(int recipeId);
+    RecipeResponse getRecipeById(long recipeId);
 }
